backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
yfinance==0.2.36
apscheduler==3.10.4
requests==2.31.0
openai==1.10.0
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3
httpx==0.27.0
feedparser==6.0.11
scikit-learn==1.4.0
google-genai>=1.0.0
psycopg2-binary==2.9.9
